Add error handlers and tidy up repo
1- Rename testing.py to LPC.py
2-Add if-else in enter_carpark and exit_carpark to handle errors in LPC.py
3 - Add Exceptions handler in numberplate.py
4- Remove some redundant .py files
