--- Verify_Installation.py
@@ -1,23 +1,0 @@
-# -----------------------------------------------------------
-#
-#
-#
-#
-#
-# -----------------------------------------------------------
-
-import os
-import numberplate
-
-print("Welcome to the carpark")
-
-entry_numberplate = numberplate.numberplate_recognition()
-#
-print(entry_numberplate)
-#def save_numberplate(numberplate_number):
-    #to be continued
-
-#save_numberplate(entry_numberplate)
-
-#if save_numberplate succeed,print something
-#else print something
