--- carpark_exit.py
@@ -1,19 +1,0 @@
-# -----------------------------------------------------------
-#
-#
-#
-#
-#
-# -----------------------------------------------------------
-
-import os
-import numberplate
-
-print("Exiting the carpark")
-
-# entry_numberplate = numberplate.read_numberplate()
-#
-# print(entry_numberplate)
-#def calulate_fee(numberplate_number):
-    #to be continued
-
